Fix config management in class hierarchy
TypeScript does not support private fields with same name in a class hierarchy, so the class local private config field must either be named differenty on every inheritance level or kept in a single place. To keep it in a sigle place and still access ass properties of the config object on very level, generics are necessary, which unecessarily blows up the code in some places. Still, the second option is the cleaner solution.