Convert component into concrete class
This component can be used for simple elements, e.g. to create an img DOM element. If we later need a base class that serves as a conceptual component that is not attached to a DOM element, this class should be renamed to Element and a new Component base class added.