Switch between normal and ad UIs by hiding/showing instead of unloading/loading
Loading and unloading UIs and thus keeping only the active UI in memory and in the DOM is the nicer solution, but it would need a lot of changes in the framework. Most importantly, subscribed event handlers on components would have to be unsubscribed, or not subscribed a second time when configuring the UI again. This would require handler functions on the class level so the can be unsubscribed or queried, and that needs a lot of refactoring. Another laternative might be to load a second UIManager for the ads UI, but then we need a manager who manages these two managers and switches between them. For now, its easier to just load both UIs and switch between them as required.