Add playback toggle overlay
This change converts the huge playback toggle button to a "normal" toggle button and adds an explicit playback toggle overlay that contains the huge playback toggle button as a child. This solves the discrepancy between the huge playback toggle button being conceptually an overlay but semantically "only" a button which stood out in the documentation as a deviation from the implicit design patterns of this framework.