Use isReady to detect the player's prepared state
Depending on when the UI is loaded, it's either the ON_READY event (in the UIContainer) or the getState function that determines the state of the UI. getState checked if a source is loaded, which must not necessarily be the case when isReady === true, e.g. when a player is setup without a source. The UI state is initialized with getState but when an ON_READY event arrives afterwards, which happens when the UI is loaded before the player setup is finished, the UI changed into the PREPARED state, no matter if a source is loaded or not. If the UI is loaded later, it only changed into the PREPARED state if a source is actually loaded. This resulted in an inconsistent behavior, depending on when the UI is loaded. Now it's the ON_READY event and isReady flag (which are equivalent) which is used to determine the UI state, resulting in a consistent behavior.