Fix AdMessageLabel being aligned to the right
This happens because in our TV skin style `ui-tv` the titlebar-top
container-wrapper children are aligning starting from the right,
so that the elements can be filled up to the left.

However, this works for normal layout since there are already multiple
elements; but in the Ads layout there is only the title, so
let's add some automatic margin to it.