Fix controls are shown when transitioning back to main content from ad
The issue is that our UI has multiple layouts existing in the DOM at
the same time (e.g. Generic layout and Ad layout).
When the Ad layout goes on, the Generic one is just hidden, but all
its element are there, in the DOM.

So we cannot prevent all hiding of the ControlBar, since we would also
prevent the hiding of the Generic layout's ControlBar.
And we do not want that; in fact, the main content ControlBat should
be hidden when we resume main content from the ad.

I fixed this by differentiating which bar should actually prevent
the hiding.

Note: hiding on AdBreakFinished is useless, as this is a bigger
concern. That's why I removed it.