Make `AdSkipButton` fixed position
Since sliding it down on controls hide has worse UX: what if the
user is pointing with the mouse or finger at the button, and the
button goes down at that same moment?

Keeping it fixed position will ensure the user does not have to
adjust their pointer position again.